agent/common/requirements-agent.txt
psutil>=5.9.0
requests>=2.28.0
pyyaml>=6.0
watchdog>=3.0.0
